// frontend/src/api/generated/model/consentDocument.ts
export type ConsentDocument = typeof ConsentDocument[keyof typeof ConsentDocument];


export const ConsentDocument = {
  resident_personal_data: 'resident_personal_data',
  guest_personal_data: 'guest_personal_data',
} as const;